backend/requirements.txt
asgiref==3.5.2
Django
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
pathspec==0.9.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.3
yamllint==1.26.3
